#
Add CommandPattern project and implement Command Pattern

This commit introduces a new "CommandPattern" project to the
SomeDesignPatterns solution, including a new project file
`CommandPattern.csproj`. The `DP.cs` file is updated to define
the `ICommand` interface and implement command classes for
creating and discounting products. A `CommandInvoker` class
is added to manage command execution, and `Program.cs` is
updated to demonstrate the usage of the Command Pattern.